add textarea onBlur for Poset mode
When the user blurs the textarea, the linear order length will be set to the maximum number.
Other changes.
Set default dragmode to "orbital" (it was "turntable"). This way, the graph can be rotated in any direction indefinitely.
Move the method selection at the bottom, adjust its width and add animation and shadow.